3D location in VisionFeature
Needs more testing. I have not been able to fully verify that everything works in visualization which I will do tonight.
There is also the edge case when a point doesn't get matched. Right now we declare that an illegal state and just assert that we never enter it. I don't think we should be doing this though.
Also I have to run so this commit isn't linted. This will be linted later.